Added a MenuManager and renamed the plugin
