--- src/io/teamelite/brush_menu/system/json/adapters/ItemStackAdapter.java
@@ -1,14 +1,0 @@
-package io.teamelite.brush_menu.system.json.adapters;
-
-/**
- * @name 		BrushMenu
- * @author 		Liam Reffell and Kieron Wiltshire
- * @contact 	http://www.mcteamelite.com/
- * @license 	MIT License
- * @description
- * 				The plugin requires Java 1.6 or higher and depends on the VoxelSniper
- * 				Bukkit plugin. It allows a user to open up an inventory interface and
- * 				select their desired VoxelSniper brush.
- */
-public class ItemStackAdapter {
-}
