+++ src/io/teamelite/menu_controller/commands/Voxel.java
@@ -1,5 +1,15 @@
-package io.teamelite.brush_menu.commands;
+package io.teamelite.menu_controller.commands;
 
+/**
+ * @name 		MenuController
+ * @author 		Liam Reffell and Kieron Wiltshire
+ * @contact 	http://www.mcteamelite.com/
+ * @license 	MIT License
+ * @description
+ * 				The plugin requires Java 1.6 or higher.
+ * 				It allows a user to open up an inventory interface and
+ * 				select their saved menu options.
+ */
 public class Voxel /* implements CommandExecutor */ {
 
 	// private Inventory inv;
