+++ src/io/teamelite/menu_controller/system/json/adapters/ItemStackAdapter.java
@@ -1,0 +1,14 @@
+package io.teamelite.menu_controller.system.json.adapters;
+
+/**
+ * @name 		MenuController
+ * @author 		Liam Reffell and Kieron Wiltshire
+ * @contact 	http://www.mcteamelite.com/
+ * @license 	MIT License
+ * @description
+ * 				The plugin requires Java 1.6 or higher.
+ * 				It allows a user to open up an inventory interface and
+ * 				select their saved menu options.
+ */
+public class ItemStackAdapter {
+}
