+++ src/io/teamelite/menu_controller/system/menu/MenuItem.java
@@ -1,16 +1,16 @@
-package io.teamelite.brush_menu.system.menu;
+package io.teamelite.menu_controller.system.menu;
 
 import org.bukkit.inventory.ItemStack;
 
 /**
- * @name 		BrushMenu
+ * @name 		MenuController
  * @author 		Liam Reffell and Kieron Wiltshire
  * @contact 	http://www.mcteamelite.com/
  * @license 	MIT License
  * @description
- * 				The plugin requires Java 1.6 or higher and depends on the VoxelSniper
- * 				Bukkit plugin. It allows a user to open up an inventory interface and
- * 				select their desired VoxelSniper brush.
+ * 				The plugin requires Java 1.6 or higher.
+ * 				It allows a user to open up an inventory interface and
+ * 				select their saved menu options.
  */
 public abstract class MenuItem {
