+++ src/io/teamelite/menu_controller/utilities/TitleManager.java
@@ -1,4 +1,4 @@
-package io.teamelite.brush_menu.utilities;
+package io.teamelite.menu_controller.utilities;
 
 import net.minecraft.server.v1_8_R2.IChatBaseComponent;
 import net.minecraft.server.v1_8_R2.PacketPlayOutTitle;
